More changes to make it public-able
Changed the default config to have XXXX's for username/password, my old
one is still in the history but I can change that at any time. Added the
"real" config to .gitignore, and created a prototyp in config.example.
Fixed up the README some. Altered install, .msmtprc and .offlineimaprc
to get things from the config file.